requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile development.in
#
amqp==5.0.5
    # via kombu
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via pytest
backcall==0.2.0
    # via ipython
beautifulsoup4==4.9.3
    # via pynliner
billiard==3.6.3.0
    # via celery
celery==5.0.5
    # via django-celery-beat
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
cryptography==3.4.6
    # via pyjwt
cssutils==1.0.2
    # via pynliner
decorator==4.4.2
    # via ipython
defusedxml==0.6.0
    # via python3-openid
django-admin-sortable2==0.7.8
    # via -r base.in
django-allauth==0.44.0
    # via -r base.in
django-appconf==1.0.4
    # via django-imagekit
django-autocomplete-light==3.8.2
    # via -r base.in
django-celery-beat==2.2.0
    # via -r base.in
django-crispy-forms==1.11.1
    # via -r base.in
django-debug-toolbar==3.2
    # via -r development.in
django-imagekit==4.0.2
    # via -r base.in
django-inlinecss==0.3.0
    # via -r base.in
django-js-asset==1.2.2
    # via django-mptt
django-model-utils==4.1.1
    # via -r base.in
django-mptt==0.11.0
    # via -r base.in
django-qr-code==2.1.0
    # via -r base.in
django-timezone-field==4.1.1
    # via django-celery-beat
django-tinymce==3.2.0
    # via -r base.in
django==3.1.6
    # via
    #   -r base.in
    #   django-admin-sortable2
    #   django-allauth
    #   django-appconf
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-inlinecss
    #   django-model-utils
    #   django-mptt
    #   django-qr-code
    #   django-timezone-field
    #   djangorestframework
djangorestframework==3.12.2
    # via -r base.in
factory-boy==3.2.0
    # via -r development.in
faker==6.1.1
    # via factory-boy
flake8==3.8.4
    # via -r development.in
future==0.18.2
    # via django-inlinecss
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
invoke==1.5.0
    # via -r base.in
ipdb==0.13.4
    # via -r development.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.20.0
    # via
    #   -r development.in
    #   ipdb
isort==5.7.0
    # via -r development.in
jedi==0.18.0
    # via ipython
kombu==5.0.2
    # via celery
mccabe==0.6.1
    # via flake8
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.9
    # via pytest
parso==0.8.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pilkit==2.0
    # via django-imagekit
pillow==8.1.0
    # via -r base.in
pip-tools==5.5.0
    # via -r development.in
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.16
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.8.6
    # via -r base.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pygments==2.7.4
    # via ipython
pyjwt[crypto]==2.0.1
    # via django-allauth
pynliner==0.8.0
    # via django-inlinecss
pyparsing==2.4.7
    # via packaging
pytest-django==4.1.0
    # via -r development.in
pytest==6.2.2
    # via
    #   -r development.in
    #   pytest-django
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.1
    # via
    #   faker
    #   python-crontab
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
redis==3.5.3
    # via -r base.in
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   django-allauth
    #   requests-oauthlib
segno==1.3.1
    # via django-qr-code
six==1.15.0
    # via
    #   click-repl
    #   django-autocomplete-light
    #   django-imagekit
    #   python-dateutil
soupsieve==2.2
    # via beautifulsoup4
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
traitlets==5.0.5
    # via ipython
urllib3==1.26.3
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
xlrd==2.0.1
    # via -r development.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
